refactor: refactor code for creating and updating person names
- The changes include updating the `createPersonName` method in
`PersonNameModel` and refactoring the code in
`VerifyingObserverSqPersistentObject`, `PatientPersistentObject`,
`SeriesPersistentObject`, and `StudyPersistentObject` classes.
- The `createPersonName` method now takes a `nameObj` parameter and
returns the created person name.
- The code in the mentioned classes that creates or updates person names
now invokes the `createPersonName` and `updatePersonNameById` methods
from the `PersonNameModel` class.
- This refactoring improves code reusability and readability.

+++ models/sql/po/instance.po.js
@@ -424,48 +424,33 @@ class VerifyingObserverSqPersistentObject {
      */
     async createName() {
         let nameItem = _.get(this.verifyingObserverSq, "0040A075.Value.0");
-        if (nameItem && Object.values(nameItem).some(v => v)) {
-            return await PersonNameModel.create({
-                alphabetic: _.get(nameItem, "Alphabetic", undefined),
-                ideographic: _.get(nameItem, "Ideographic", undefined),
-                phonetic: _.get(nameItem, "Phonetic", undefined)
-            });
-        }
-        return undefined;
+        return await PersonNameModel.createPersonName(nameItem);
     }
 
     async updateName() {
         let verifyingObserverSq = await this.getExistItem();
+        /** @type { PersonNameModel | undefined }  */
         let name = await verifyingObserverSq.getPersonName();
         let nameItem = _.get(this.verifyingObserverSq, "0040A075.Value.0");
         
         if (name) {
-            if (nameItem && Object.values(nameItem).some(v => v)) {
-                await PersonNameModel.update({
-                    alphabetic: _.get(nameItem, "Alphabetic", undefined),
-                    ideographic: _.get(nameItem, "Ideographic", undefined),
-                    phonetic: _.get(nameItem, "Phonetic", undefined)
+            let updatedName = await PersonNameModel.updatePersonNameById(nameItem, name.getDataValue("id"));
+            if (!updatedName) {
+                await VerifyIngObserverSqModel.update({
+                    x0040A075: null
                 }, {
+                    where: {
+                        SOPInstanceUID: this.instance.dataValues.x00080018
+                    }
+                });
+                await PersonNameModel.destroy({
                     where: {
                         id: name.dataValues.id
                     }
                 });
+            } else {
                 return name;
             }
- 
-            await VerifyIngObserverSqModel.update({
-                x0040A075: null
-            }, {
-                where: {
-                    SOPInstanceUID: this.instance.dataValues.x00080018
-                }
-            });
-            await PersonNameModel.destroy({
-                where: {
-                    id: name.dataValues.id
-                }
-            });
-
         } else {
             return await this.createName();
         }
